✅ test: myBatis 테스트 코드 작성
- 테스트 시 문제였던 저장 시 글 객체의 id가 저장 안되는 문제 해결 (useGeneratedKeys 직접 작성)

+++ src/main/java/devholic22/board/repository/mybatis/MyBatisBoardRepository.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Repository
 @RequiredArgsConstructor
@@ -16,8 +17,9 @@ public class MyBatisBoardRepository implements BoardRepository {
     private final BoardMapper boardMapper;
 
     @Override
-    public void save(Board board) {
+    public Board save(Board board) {
         boardMapper.save(board);
+        return board;
     }
 
     @Override
@@ -41,12 +43,18 @@ public class MyBatisBoardRepository implements BoardRepository {
     }
 
     @Override
-    public void update(Integer id, BoardDto boardDto) {
+    public void update(Integer id, BoardDto boardDto) throws NullPointerException {
+        if (boardMapper.findById(id) == null) {
+            throw new NullPointerException();
+        }
         boardMapper.update(id, boardDto);
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Integer id) throws NoSuchElementException {
+        if (boardMapper.findById(id) == null) {
+            throw new NoSuchElementException();
+        }
         boardMapper.delete(id);
     }
